run beforeInsert, afterInsert etc. on docs inserted across locales, plus eslint compliance and enforcement for this project
