+++ lib/api.js
@@ -12,7 +12,7 @@ var diff = require('jsondiffpatch').create({
     return obj._id || obj.id || JSON.stringify(obj);
   },
   textDiff: {
-    // Don't try to diff text — replace it. Otherwise
+    // Don't try to diff text, replace it. Otherwise
     // patches are never applicable across locales
     minLength: 1000000000
   },
@@ -24,7 +24,7 @@ var diff = require('jsondiffpatch').create({
 });
 
 module.exports = function(self, options) {
-  
+
   // Resolve relationships between this doc and other docs, which need to be
   // mapped to the appropriate doc in the new locale, via the workflowGuid
   // property of each doc.
@@ -58,7 +58,7 @@ module.exports = function(self, options) {
       var workflowGuids;
       var secondLocaleIds;
       var originalIds;
-      
+
       return async.eachSeries(joins, function(join, callback) {
         if (!self.includeType(join.withType)) {
           // Type joined with is not subject to workflow
@@ -74,7 +74,7 @@ module.exports = function(self, options) {
           remapDocs();
           return callback(null);
         });
-        
+
         function findWorkflowGuids(callback) {
           if (join.field.type === 'joinByOne') {
             return self.apos.docs.db.find({ _id: { $in: [ join.doc[join.field.idField] ] } }, { workflowGuid: 1 }).toArray(function(err, docs) {
@@ -140,18 +140,18 @@ module.exports = function(self, options) {
         }
 
       }, callback);
-                
-    } 
- 
+
+    }
+
   };
-  
+
   // Invokes resolveRelationships for all docs that have the
   // workflowResolveDeferred: true property, and removes the property.
   //
   // This is invoked by migration tasks as a second step after copying the docs
   // across locales, so that all of the joined documents actually exist in
   // the other locales before the resolver looks for them.
-  
+
   self.resolveDeferredRelationships = function(callback) {
     var req = self.apos.tasks.getReq();
     return self.apos.migrations.eachDoc({ workflowResolveDeferred: true }, function(doc, callback) {
@@ -162,11 +162,11 @@ module.exports = function(self, options) {
         resolveRelationships,
         update
       ], callback);
-      
+
       function resolveRelationships(callback) {
         return self.resolveRelationships(req, doc, doc.workflowLocale, callback);
       }
-      
+
       function update(callback) {
         delete doc.workflowResolveDeferred;
         return self.apos.docs.db.update({ _id: doc._id }, doc, callback);
@@ -174,7 +174,7 @@ module.exports = function(self, options) {
     }, callback);
   };
 
-  // You probably want `commitLatest`. 
+  // You probably want `commitLatest`.
   //
   // Commit a doc from one locale to another. `from` and `to` should
   // be the doc as found in each of the locales. The callback receives
@@ -220,7 +220,7 @@ module.exports = function(self, options) {
       }, callback);
     }
   };
-  
+
   // Fetch the draft version of a doc, whose id is `id`, and also the live version of the same
   // doc. On success, deliver `(null, draft, live)` to the callback.
   //
@@ -234,7 +234,7 @@ module.exports = function(self, options) {
   //
   // By default, areas and joins are fully loaded. To disable that,
   // set `options.areas` and `options.joins` to `false`.
-  
+
   self.getDraftAndLive = function(req, id, options, callback) {
     var draft;
     var live;
@@ -259,7 +259,7 @@ module.exports = function(self, options) {
         if ((options.permission !== false) && !_one._edit) {
           return callback('notfound');
         }
-        if (_one.workflowLocale && _one.workflowLocale.match(/\-draft$/)) {
+        if (_one.workflowLocale && _one.workflowLocale.match(/-draft$/)) {
           draft = _one;
         } else {
           live = _one;
@@ -286,12 +286,12 @@ module.exports = function(self, options) {
         }
         return callback(null);
       });
-    }      
+    }
   };
-          
+
   // Delete properties of doc that are not considered
   // relevant to workflow.
-  
+
   self.deleteExcludedProperties = function(doc) {
     // console.log('before delete:');
     // console.log(JSON.stringify(doc, null, '  '));
@@ -303,7 +303,7 @@ module.exports = function(self, options) {
     // console.log('after delete:');
     // console.log(JSON.stringify(doc, null, '  '));
   };
-      
+
   // Copy properties that are included in workflow from the doc `from`
   // to the doc `to`. TODO: this method does not yet address copying
   // modified attachments to make sure an edit to one locale does not alter
@@ -320,7 +320,7 @@ module.exports = function(self, options) {
     // * Not damage everything else
     //
     // So we copy all schema properties and top-level areas not excluded.
-    
+
     _.each(from, function(val, key) {
       if (self.includeProperty(key)) {
         to[key] = val;
@@ -329,10 +329,10 @@ module.exports = function(self, options) {
 
     // TODO deal with copying attachments rather than referencing
     // the same file, however take care not to do it if there is no change
-    
+
     return setImmediate(callback);
   };
-  
+
   // Delete top-level area properties of `to` that do not exist
   // in `from`, unless they are explicitly excluded from workflow.
   //
@@ -356,7 +356,7 @@ module.exports = function(self, options) {
     return setImmediate(callback);
   };
 
-  // Returns true if this top level doc property should be included 
+  // Returns true if this top level doc property should be included
   // when committing changes from draft to live
   self.includeProperty = function(prop) {
     if (prop.match(/^_?workflow/)) {
@@ -367,7 +367,7 @@ module.exports = function(self, options) {
     }
     return true;
   };
-      
+
   // The callback will receive basic information about all docs editable by the
   // current user that are awaiting approval to merge from draft to live.
   //
@@ -394,7 +394,7 @@ module.exports = function(self, options) {
     };
     return self.apos.docs.find(req, criteria, self.getSubmittedProjection()).sort({ 'workflowSubmitted.when': -1 }).trash(null).published(null).toArray(callback);
   };
-      
+
   // Returns the projection to be used when fetching submitted docs to generate
   // a list of docs requiring approval. Should be enough to generate permalinks.
 
@@ -409,16 +409,16 @@ module.exports = function(self, options) {
       workflowSubmitted: 1
     };
   };
-  
+
   // Returns a cursor to find docs even if they are in the trash or
   // unpublished, without regard to locale (except for permissions checks) unless
   // the locale argument is present. The locale argument may be completely
   // omitted. Areas and joins are not loaded by default.
-  
+
   self.findDocs = function(req, criteria, locale) {
     return self.apos.docs.find(req, criteria).trash(null).published(null).workflowLocale(locale || null).areas(false).joins(false);
   };
-      
+
   // Fetch a draft doc along with all of the past commits in which it is the source ("fromId").
   //
   // On success, invokes callback with `(null, doc, commits)` where `commits` is an array.
@@ -476,7 +476,7 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function getDoc(callback) {
       return self.findDocs(req, { _id: commit.fromId }).areas(true).joins(true).toObject(function(err, _doc) {
         if (err) {
@@ -491,7 +491,7 @@ module.exports = function(self, options) {
     }
 
   };
-      
+
   // Decide whether a doc type is subject to workflow as documented for the module options.
 
   self.includeType = function(type) {
@@ -518,7 +518,7 @@ module.exports = function(self, options) {
   // typically (live vs. draft), deliver `(null, fields)`
   // to the callback, where `fields` is an array of
   // schema field names that have been modified.
-  
+
   self.getModifiedFields = function(req, before, after, callback) {
     return self.resolveRelationships(req, before, after.workflowLocale, function(err) {
       if (err) {
@@ -560,9 +560,9 @@ module.exports = function(self, options) {
       return callback(null, modifiedFields);
     });
   };
-      
+
   self.draftify = function(locale) {
-    if (locale.match(/\-draft$/)) {
+    if (locale.match(/-draft$/)) {
       return locale;
     } else {
       return locale + '-draft';
@@ -570,13 +570,13 @@ module.exports = function(self, options) {
   };
 
   self.liveify = function(locale) {
-    return locale.replace(/\-draft$/, '');
+    return locale.replace(/-draft$/, '');
   };
 
   // Return an object ready for use as the `workflowSubmitted`
   // property of a doc. If `data` is provided, any properties of
   // `data` are added to the object.
-  
+
   self.getWorkflowSubmittedProperty = function(req, data) {
     return _.assign({
       username: req.user.username,
@@ -619,7 +619,6 @@ module.exports = function(self, options) {
       }
       relatedModified = filter(relatedModified);
       relatedUnmodified = filter(relatedUnmodified);
-      var byId = {};
       var committable = [];
       _.each(relatedModified.concat(relatedUnmodified), function(doc) {
         if (_.has(liveVersions, doc._id)) {
@@ -663,7 +662,7 @@ module.exports = function(self, options) {
         return callback(err);
       });
     }
-      
+
     function diff(callback) {
       var ids = _.map(related, function(item) {
         return item._id || (item.item && item.item._id);
@@ -704,7 +703,7 @@ module.exports = function(self, options) {
           }
           return callback(null);
         });
-                   
+
         function getDraftAndLive(callback) {
           if (_.has(complete, id)) {
             draft = complete[id];
@@ -721,14 +720,14 @@ module.exports = function(self, options) {
             return callback(null);
           });
         }
-        
+
         function resolveRelationships(callback) {
-          self.resolveRelationships(req, draft, live.workflowLocale, callback); 
+          self.resolveRelationships(req, draft, live.workflowLocale, callback);
         }
       }, callback);
     }
   };
-  
+
   self.getRelated = function(docs, callback) {
     var result = [];
     _.each(docs, function(draft) {
@@ -746,7 +745,7 @@ module.exports = function(self, options) {
     });
     return callback(null, result);
   };
-  
+
   // Fetch joins, load areas, etc. on doc objects that came out of the
   // commits collection. Used for previewing.
 
@@ -774,9 +773,9 @@ module.exports = function(self, options) {
     };
     // Are we interested in draft locales, or live locales?
     if (draft) {
-      criteria.workflowLocale = /\-draft$/;
+      criteria.workflowLocale = /-draft$/;
     } else {
-      criteria.workflowLocale = { $not: /\-draft$/ };
+      criteria.workflowLocale = { $not: /-draft$/ };
     }
     return self.apos.docs.find(req, criteria, self.getContextProjection()).workflowLocale(null).published(null).addUrls(false).toArray(function(err, docs) {
       if (err) {
@@ -800,13 +799,13 @@ module.exports = function(self, options) {
       return callback(null, localizations);
     });
   };
-  
+
   // Returns the doc that is logically thought of as
   // the "context" for the current page rendering, i.e.
   // a piece if on a show page, a page otherwise
 
   self.getContext = function(req) {
-    return req.data.piece || req.data.page;      
+    return req.data.piece || req.data.page;
   };
 
   // Render the contextual action buttons — draft/live, submit and commit.
@@ -827,7 +826,7 @@ module.exports = function(self, options) {
   //
   // On success the callback receives `(null, _id)` where `_id` is
   // the unique identifier for this specific commit in the collection.
-  
+
   self.insertCommit = function(req, from, to, callback) {
     var _id = self.apos.utils.generateId();
     return self.db.insert({
@@ -847,7 +846,7 @@ module.exports = function(self, options) {
       return callback(null, _id);
     });
   };
-  
+
   self.enableHelpers = function() {
     self.addHelpers({
       localizations: function() {
@@ -908,7 +907,7 @@ module.exports = function(self, options) {
       return self.apos.permissions.can(_req, 'edit-' + doc.type);
     });
   };
-  
+
   // As much as is possible, apply the changes that occurred
   // between `to` (an older revision of the doc) and `from`
   // (a newer revision, FROM which the changes come)
@@ -947,7 +946,7 @@ module.exports = function(self, options) {
       description = draft;
       draft = null;
     }
-    
+
     // Step 0: make sure areas exist so we don't wind up with patches that remove them
     // or nowhere to add them
 
@@ -992,9 +991,9 @@ module.exports = function(self, options) {
     });
 
     var draftObjects = draft && getObjects(draft);
-                
+
     // Step 2: iterate over those objects, patching directly as appropriate
-    
+
     // 2a: remove deleted things. "to" is the old version, PRIOR to the commit.
     // Anything present there and absent in "from" was therefore removed DURING the commit
     _.each(toObjects.objects, function(value) {
@@ -1015,7 +1014,7 @@ module.exports = function(self, options) {
         deleteObject(to, toObjects, value);
       }
     });
-    
+
     // 2b: moved things and brand new things. For these we start with parent
     // objects, so that there is a context for patching child objects in
     // the right location later
@@ -1044,7 +1043,7 @@ module.exports = function(self, options) {
         recomputeDotPaths(from, fromObjects);
       }
     });
-    
+
     // 2c: updated things. Here we start with child objects, so
     // that we can figure out parents have not changed
 
@@ -1055,7 +1054,7 @@ module.exports = function(self, options) {
         // We already dealt with additions in 2b
         return;
       }
-      // Modified. 
+      // Modified.
       if (!_.isEqual(value, toObjects.byId[value._id])) {
         if (description) {
           // console.log(toObjects.dotPaths[value._id]);
@@ -1073,15 +1072,15 @@ module.exports = function(self, options) {
       recomputeDotPaths(from, fromObjects);
       if (deleteAfter) {
         deleteObject(from, fromObjects, value);
-        deleteObject(to, toObjects, value);              
+        deleteObject(to, toObjects, value);
       }
     });
-    
+
     // Step 3: remove any remaining _id objects in commit.from and commit.to
     // so jsondiffpatch doesn't consider them
     purgeObjects(from, fromObjects);
     purgeObjects(to, toObjects);
-    
+
     // Step 4: patch with jsondiff for everything that doesn't have an _id
 
     if (description) {
@@ -1096,14 +1095,14 @@ module.exports = function(self, options) {
     } catch (e) {
       return callback(null, 'Content was too different');
     }
-  
+
     function getObjects(doc) {
       var objects = [];
       var dotPaths = {};
       var dots = {};
-      
+
       self.apos.docs.walk(doc, function(doc, key, value, dotPath, ancestors) {
-        if (value && (typeof(value) === 'object') && value._id) {
+        if (value && (typeof (value) === 'object') && value._id) {
           objects.push(value);
           dotPaths[value._id] = dotPath;
           dots[value._id] = 0;
@@ -1114,9 +1113,9 @@ module.exports = function(self, options) {
           }
         }
       });
-            
-      topDownObjects = _.clone(objects);
-           
+
+      var topDownObjects = _.clone(objects);
+
       objects.sort(function(a, b) {
         if (dots[a._id] > dots[b._id]) {
           return -1;
@@ -1126,7 +1125,7 @@ module.exports = function(self, options) {
           return 0;
         }
       });
-      
+
       topDownObjects.sort(function(a, b) {
         if (dots[a._id] > dots[b._id]) {
           return 1;
@@ -1146,21 +1145,21 @@ module.exports = function(self, options) {
       };
 
     }
-    
+
     function recomputeDotPaths(doc, objects) {
       var dotPaths = {};
       self.apos.docs.walk(doc, function(doc, key, value, dotPath, ancestors) {
-        if (value && (typeof(value) === 'object') && value._id) {
+        if (value && (typeof (value) === 'object') && value._id) {
           dotPaths[value._id] = dotPath;
         }
       });
       objects.dotPaths = dotPaths;
     }
-               
+
     function getObject(context, objects, dotPath) {
       return deep(context, dotPath);
     }
-    
+
     function deleteObject(context, objects, value) {
       var dotPath = objects.dotPaths[value._id];
       if (!dotPath) {
@@ -1184,7 +1183,7 @@ module.exports = function(self, options) {
         // }
       }
     }
-    
+
     function insertObjectAfter(context, objects, afterId, value) {
       var afterDotPath = objects.dotPaths[afterId];
       var stem = self.getStem(afterDotPath);
@@ -1205,8 +1204,8 @@ module.exports = function(self, options) {
       recomputeDotPaths(context, objects);
       // In case this is an insertion
       objects.byId[value._id] = value;
-    }          
-                
+    }
+
     function appendObject(context, objects, path, object) {
       var array = deep(context, path);
       if (!Array.isArray(array)) {
@@ -1217,7 +1216,7 @@ module.exports = function(self, options) {
       objects.byId[object._id] = object;
       return true;
     }
-    
+
     function purgeObjects(context, objects) {
       _.each(objects.objects, function(object) {
         deleteObject(context, objects, object);
@@ -1239,11 +1238,13 @@ module.exports = function(self, options) {
         }
       }
     }
-    
+
     function moved(fromDotPath, value, isNew) {
-      var fromIndex = parseInt(_.last(fromDotPath.split('.')));   
+      var fromIndex = parseInt(_.last(fromDotPath.split('.')));
       var afterId;
       var i;
+      var subPath;
+      var obj;
       // We're moving it, not modifying it, so go with the
       // existing value in the draft if there is one
       var toMove = (draftObjects && draftObjects.byId[value._id]) || value;
@@ -1252,41 +1253,40 @@ module.exports = function(self, options) {
         // `from` and `draft`, and position the moved object
         // after it
         for (i = fromIndex - 1; (i >= 0); i--) {
-          var subPath = fromDotPath.split('.');
+          subPath = fromDotPath.split('.');
           subPath.pop();
           subPath.push(i);
-          var obj = getObject(from, fromObjects, subPath.join('.'));
-          var afterId = obj._id;
+          obj = getObject(from, fromObjects, subPath.join('.'));
+          afterId = obj._id;
           if (_.has((draftObjects || toObjects).byId, afterId)) {
             if (draft) {
               deleteObject(draft, draftObjects, toMove);
               insertObjectAfter(draft, draftObjects, afterId, toMove);
             } else {
               description.push({
-                change: isNew ? 'added': 'moved',
+                change: isNew ? 'added' : 'moved',
                 value: value,
                 _id: value._id
               });
             }
             return;
           }
-        }  
+        }
       } else {
         // Looking for something to insert "after" won't work
         // for the very first widget in the area. So find something
         // in common that follows it and insert before that
         i = 1;
         while (true) {
-          var subPath = fromDotPath.split('.');
+          subPath = fromDotPath.split('.');
           subPath.pop();
           subPath.push(i);
-          var obj = getObject(from, fromObjects, subPath.join('.'));
+          obj = getObject(from, fromObjects, subPath.join('.'));
           if (!obj) {
             // No more to look at
             break;
           }
           var beforeId = obj._id;
-          var toMove;
           if (_.has((draftObjects || toObjects).byId, beforeId)) {
             if (draft) {
               // We're moving it, not modifying it, so go with the
@@ -1296,7 +1296,7 @@ module.exports = function(self, options) {
               insertObjectBefore(draft, draftObjects, beforeId, toMove);
             } else {
               description.push({
-                change: isNew ? 'added': 'moved',
+                change: isNew ? 'added' : 'moved',
                 value: value,
                 _id: value._id
               });
@@ -1308,25 +1308,25 @@ module.exports = function(self, options) {
             return;
           }
           i++;
-        }  
+        }
       }
       // Fallback when we can't find any peers in
       // `draft`: append it
       if (description) {
         description.push({
-          change: isNew ? 'added': 'moved',
+          change: isNew ? 'added' : 'moved',
           value: value,
           _id: value._id
         });
         deleteObject(from, fromObjects, value);
-        deleteObject(to, toObjects, value);              
+        deleteObject(to, toObjects, value);
       } else {
         deleteObject(draft, draftObjects, value);
         if (appendObject(draft, draftObjects, fromDotPath.replace(/\.\d+$/, ''), toMove)) {
           // So we know the difference no longer exists when examining
           // a parent object
           deleteObject(from, fromObjects, value);
-          deleteObject(to, toObjects, value);              
+          deleteObject(to, toObjects, value);
         } else {
           // append failed, probably because the parent object
           // doesn't exist yet in the draft. So don't purge or
@@ -1374,7 +1374,7 @@ module.exports = function(self, options) {
     if ((!doc.workflowGuid) || (!self.includeType(doc.type))) {
       // doc is not subject to workflow. For developer convenience, allow this
       // method to be used in that case too; it just updates the single doc
-      return self.apos.docs.db.update({ $and: [ { _id: doc._id }, apos.permissions.criteria(req, 'edit-' + doc.type) ] }, { $set: set }, callback);
+      return self.apos.docs.db.update({ $and: [ { _id: doc._id }, self.apos.permissions.criteria(req, 'edit-' + doc.type) ] }, { $set: set }, callback);
     }
 
     return self.apos.docs.db.update(
@@ -1384,7 +1384,7 @@ module.exports = function(self, options) {
       callback
     );
   };
-  
+
   // See `setPropertiesAcrossLocales`.
   //
   // Return a MongoDB criteria object selecting the given doc across multiple
@@ -1425,7 +1425,7 @@ module.exports = function(self, options) {
     };
     return criteria;
   };
-  
+
   // Commits the current draft of the given doc id
   // to the live locale for that document. `id`
   // may be either the draft or the live id.
@@ -1434,13 +1434,13 @@ module.exports = function(self, options) {
   //
   // The `id` argument is sanitized, so it is safe
   // to pass user input directly.
-  
+
   self.commitLatest = function(req, id, callback) {
     if (!req.user) {
       // confusion to the enemy
       return callback('error');
     }
-    var id = self.apos.launder.id(id);
+    id = self.apos.launder.id(id);
     var draft, live, commitId;
     return async.series({
       getDraftAndLive,
@@ -1467,8 +1467,8 @@ module.exports = function(self, options) {
         return callback(err);
       });
     }
-  };  
-  
+  };
+
   // Export the given commit id (NOT doc id) to the given locales.
   // On success the callback receives `(null, result)`
   // where `result` is an object with `success` and `errors` properties.
@@ -1489,18 +1489,18 @@ module.exports = function(self, options) {
       // Confusion to the enemy
       return callback('error');
     }
-    
-    var id = self.apos.launder.id(id);
+
+    id = self.apos.launder.id(id);
 
     if (Array.isArray(locales)) {
       locales = _.filter(locales, function(locale) {
-        return ((typeof(locale) === 'string') && (_.has(self.locales, locale)));
+        return ((typeof (locale) === 'string') && (_.has(self.locales, locale)));
       });
     }
     locales = _.map(locales, function(locale) {
       return self.draftify(locale);
     });
-    
+
     var success = [];
     var errors = [];
     var commit;
@@ -1543,7 +1543,7 @@ module.exports = function(self, options) {
 
         from = _.cloneDeep(commit.from);
         to = _.cloneDeep(commit.to);
-        
+
         return async.series([ getDraft, resolveToSource, applyPatch, resolveToDestination, update ], callback);
 
         function getDraft(callback) {
@@ -1578,7 +1578,7 @@ module.exports = function(self, options) {
             errors.push({ locale: self.liveify(locale), message: 'not found, run task' });
             return callback(null);
           }
-          
+
           return self.applyPatch(to, from, draft, function(err) {
             if (err) {
               errors.push({ locale: self.liveify(locale), message: 'Some or all content was too different' });
@@ -1603,8 +1603,8 @@ module.exports = function(self, options) {
         }
       }, callback);
     }
-  };  
-  
+  };
+
   // Force export the doc with the given id to the given locales.
   // the id and locales arguments are sanitized, so it is safe
   // to pass user input directly.
@@ -1617,19 +1617,19 @@ module.exports = function(self, options) {
   // overall error because it can be a normal situation and
   // does not indicate a systemic problem (TODO: is this
   // really true for this method in the same way it is true for export?)
-  
+
   self.forceExport = function(req, id, locales, callback) {
     if (!req.user) {
       // Confusion to the enemy
       return callback('error');
     }
-    var id = self.apos.launder.id(id);
+    id = self.apos.launder.id(id);
     var success = [];
     var errors = [];
-    var drafts, original;
+    var original;
     if (Array.isArray(locales)) {
       locales = _.filter(locales, function(locale) {
-        return ((typeof(locale) === 'string') && (_.has(self.locales, locale)));
+        return ((typeof (locale) === 'string') && (_.has(self.locales, locale)));
       });
     }
     locales = _.map(locales, function(locale) {
@@ -1670,7 +1670,7 @@ module.exports = function(self, options) {
       return async.eachSeries(locales, function(locale, callback) {
 
         var resolvedOriginal, draft;
-        
+
         // Our own modifiable copy to safely pass to `resolveToDestination`
         resolvedOriginal = _.cloneDeep(original);
 
@@ -1703,20 +1703,20 @@ module.exports = function(self, options) {
           }
 
           self.deleteExcludedProperties(resolvedOriginal);
-          
+
           _.assign(draft, resolvedOriginal);
-          
+
           return callback(null);
-                      
+
         }
 
         function update(callback) {
           success.push(self.liveify(draft.workflowLocale));
-          return self.apos.docs.update(req, draft, callback); 
+          return self.apos.docs.update(req, draft, callback);
         }
 
       }, callback);
-    }     
+    }
   };
 
 };
