+++ lib/browser.js
@@ -12,5 +12,5 @@ module.exports = function(self, options) {
     self.pushAsset('script', 'batch-export-modal', { when: 'user' });
     self.pushAsset('script', 'batch-force-export-modal', { when: 'user' });
     self.pushAsset('stylesheet', 'user', { when: 'user' });
-  };  
+  };
 };
