+++ lib/callAll.js
@@ -11,11 +11,11 @@ module.exports = function(self, options) {
     if (!self.includeType(doc.type)) {
       return;
     }
-        
+
     self.ensureWorkflowLocale(req, doc);
 
     self.ensurePageSlugPrefix(doc);
-    
+
   };
 
   // Every time a doc is saved, check whether its type is included in workflow. If it is,
@@ -28,7 +28,7 @@ module.exports = function(self, options) {
   self.docAfterSave = function(req, doc, options, callback) {
 
     var missingLocales;
-    
+
     if (doc._workflowPropagating) {
       // Recursion guard
       return callback(null);
@@ -101,7 +101,7 @@ module.exports = function(self, options) {
           // If it's a draft let it through as live, otherwise
           // start it in the trash. This is the least confusing behavior
           // for the add-missing-locales task
-          if (!_doc.workflowLocale.match(/\-draft$/)) {
+          if (!_doc.workflowLocale.match(/-draft$/)) {
             _doc.trash = true;
           }
         } else if (!options.workflowMissingLocalesLive) {
@@ -119,17 +119,39 @@ module.exports = function(self, options) {
           }
           return self.resolveRelationships(req, _doc, _doc.workflowLocale, callback);
         }
-        
+
         function insert(callback) {
-          // TODO: copy attachments so they are not directly shared resulting in cross-locale modification
-          return self.apos.docs.insert(req, _doc, { permissions: false }, function(err) {
-            return callback(err);
-          });
+          // This is tricky: for pieces, we need the beforeInsert/afterInsert etc.
+          // callbacks to run. Whereas for pages, not all of those exist, and those
+          // that do are methods of the pages module, not the manager for a
+          // particular page type.
+          var manager = self.apos.docs.getManager(_doc.type);
+          if (self.apos.instanceOf(manager, 'apostrophe-custom-pages')) {
+            // All page type managers extend apostrophe-custom-pages (eventually)
+            return async.series([ beforeInsert, beforeSave, insertDoc ], callback);
+          } else if (manager.insert) {
+            // A piece, or something else with a direct insert method;
+            // simple
+            return manager.insert(req, _doc, { permissions: false }, callback);
+          } else {
+            // Something Else. Currently, nothing in this category, but
+            // inserting via docs.insert is a good fallback
+            return insertDoc(callback);
+          }
+          function beforeInsert(callback) {
+            return self.apos.pages.beforeInsert(req, _doc, options, callback);
+          }
+          function beforeSave(callback) {
+            return self.apos.pages.beforeSave(req, _doc, options, callback);
+          }
+          function insertDoc(callback) {
+            return self.apos.docs.insert(req, _doc, options, callback);
+          }
         }
 
       }, callback);
     }
-    
+
     function permissionsAcrossLocales(callback) {
       // If I can edit a specific page in ch-fr, I can also edit that same page in gb-en,
       // PROVIDED THAT I can edit pages in gb-en at all (we have locale-specific
@@ -156,10 +178,10 @@ module.exports = function(self, options) {
         multi: true
       }, callback);
     }
-  }; 
+  };
 
   self.pageBeforeSend = function(req, callback) {
-    
+
     // If looking at a live locale, disable inline editing
     if (req.user && (req.session.workflowMode === 'live')) {
       req.disableEditing = true;
@@ -182,7 +204,7 @@ module.exports = function(self, options) {
       getLocalizations,
       userOnly
     ], callback);
-    
+
     function getLocalizations(callback) {
       if (!(req.data.workflow.context && req.data.workflow.context.workflowGuid)) {
         return callback(null);
@@ -195,8 +217,11 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function userOnly(callback) {
+
+      var id;
+
       // If we're not logged in, this is as far as we need to go
       if (!req.user) {
         return callback(null);
@@ -208,7 +233,7 @@ module.exports = function(self, options) {
       self.pushCreateSingleton(req);
       if (req.query.workflowPreview) {
         req.disableEditing = true;
-        var id = self.apos.launder.id(req.query.workflowPreview);
+        id = self.apos.launder.id(req.query.workflowPreview);
         self.apos.templates.addBodyClass(req, 'apos-workflow-preview-page');
         req.browserCall('apos.modules["apostrophe-workflow"].enablePreviewIframe({ id: ? })', id);
       }
@@ -219,10 +244,10 @@ module.exports = function(self, options) {
       if (!req.query.workflowReview) {
         return callback(null);
       }
-      
+
       req.disableEditing = true;
       // A commit id, not a doc id
-      var id = self.apos.launder.id(req.query.workflowReview);
+      id = self.apos.launder.id(req.query.workflowReview);
       self.apos.templates.addBodyClass(req, 'apos-workflow-preview-page');
 
       var commit;
@@ -230,7 +255,7 @@ module.exports = function(self, options) {
 
       return async.series([
         findDocAndCommit,
-        after        
+        after
       ], function(err) {
         if (err) {
           return callback(err);
@@ -247,11 +272,11 @@ module.exports = function(self, options) {
           commit = _commit;
           // Walk recursively through req.data looking for instances of the doc of interest.
           // Working in place, modify them to be copies of commit.from, which will be
-          // an older version of the doc. Since we're working in place, make 
+          // an older version of the doc. Since we're working in place, make
           // an array of these to pass to after() later for joining
           contexts = [];
           self.apos.docs.walk(req.data, function(o, k, v, dotPath) {
-            if (v && (typeof(v) === 'object')) {
+            if (v && (typeof (v) === 'object')) {
               if (v._id === commit.fromId) {
                 _.each(_.keys(v), function(key) {
                   delete v[key];
@@ -264,19 +289,19 @@ module.exports = function(self, options) {
           return callback(null);
         });
       }
-      
+
       function after(callback) {
         return self.after(req, contexts, callback);
       }
 
     }
   };
-  
+
   self.loginDeserialize = function(user) {
     user._permissionsLocales = {};
     _.each(user._groups, function(group) {
       _.merge(user._permissionsLocales, group.permissionsLocales || {});
     });
   };
-   
+
 };
