+++ lib/helpers.js
@@ -6,8 +6,7 @@ module.exports = function(self, options) {
       if (self.locales[locale] && self.locales[locale].lang) {
         return self.locales[locale].lang;
       }
-      return locale.replace(/[\-\_]\w+$/, '');
+      return locale.replace(/[-_]\w+$/, '');
     }
-  }); 
+  });
 };
-
