+++ lib/implementation.js
@@ -13,7 +13,7 @@ module.exports = function(self, options) {
         label: 'Workflow'
       }
     ];
-    
+
     self.locales = {};
     flattenLocales(self.nestedLocales);
     if (_.keys(self.locales).length > 1) {
@@ -21,7 +21,7 @@ module.exports = function(self, options) {
     }
 
     addDraftLocales();
-    
+
     function flattenLocales(locales) {
       _.each(locales, function(locale) {
         self.locales[locale.name] = locale;
@@ -30,7 +30,7 @@ module.exports = function(self, options) {
         }
       });
     }
-    
+
     function addDraftLocales() {
       var newLocales = {};
       _.each(self.locales, function(locale, name) {
@@ -43,11 +43,11 @@ module.exports = function(self, options) {
       });
       self.locales = newLocales;
     }
-    
+
     self.defaultLocale = options.defaultLocale || 'default';
-    
+
   };
-  
+
   self.composeOptions = function() {
     self.baseExcludeProperties = options.baseExcludeProperties || [
       '_id',
@@ -61,7 +61,7 @@ module.exports = function(self, options) {
       'highSearchWords',
       'searchSummary',
       'parked',
-      
+
       // Permissions are propagated across all locales by docAfterSave,
       // so it is redundant and inappropriate to include them in workflow
       'docPermissions',
@@ -83,11 +83,11 @@ module.exports = function(self, options) {
       'editGroupsRemovedIds',
       'advisoryLock'
     ];
-    
+
     // Attachment fields themselves are not directly localized (they are not docs)
     self.excludeActions = (self.options.excludeActions || []).concat(self.options.baseExcludeActions || [ 'admin', 'edit-attachment' ]);
-    
-    self.includeVerbs = (self.options.includeVerbs || []).concat(self.options.baseIncludeVerbs  || [ 'admin', 'edit' ]);
+
+    self.includeVerbs = (self.options.includeVerbs || []).concat(self.options.baseIncludeVerbs || [ 'admin', 'edit' ]);
 
     // Localizing users and groups raises serious security questions. If they have a public representation,
     // make a new doc type and join to it
@@ -96,14 +96,14 @@ module.exports = function(self, options) {
     // In 2.x, workflow applies to every property not explicitly excluded,
     // so configuration is simpler (localization will refine this though)
     self.excludeProperties = self.baseExcludeProperties.concat(options.excludeProperties || []);
-    
+
     self.includeTypes = options.includeTypes || false;
-    self.excludeTypes = self.baseExcludeTypes.concat(options.excludeTypes || []);      
+    self.excludeTypes = self.baseExcludeTypes.concat(options.excludeTypes || []);
 
     if (self.options.hostnames) {
       self.hostnames = self.options.hostnames;
     }
-        
+
     if (self.options.prefixes === true) {
       self.prefixes = {};
       _.each(self.locales, function(locale, name) {
@@ -119,7 +119,7 @@ module.exports = function(self, options) {
         }
         prefix = prefix || '';
         prefix = prefix.toString();
-        if (!prefix.match(/^\/[^\/]+$/)) {
+        if (!prefix.match(/^\/[^/]+$/)) {
           throw new Error('apostrophe-workflow: prefixes option: prefix ' + prefix + ' is invalid. If present it must be / followed by non-slash characters only');
         }
       });
@@ -131,10 +131,10 @@ module.exports = function(self, options) {
 
   // Ensure the given doc has a `workflowLocale` property; if not
   // supply it from `req.locale`, also creating a new `workflowGuid`
-  // and invoking `ensureWorkflowLocaleForPathIndex`. 
+  // and invoking `ensureWorkflowLocaleForPathIndex`.
   //
   // If the locale's type is not included in workflow, nothing happens.
-  
+
   self.ensureWorkflowLocale = function(req, doc) {
     if (!self.includeType(doc.type)) {
       return;
@@ -143,7 +143,7 @@ module.exports = function(self, options) {
     // locale, or the default locale
     if (!doc.workflowLocale) {
       doc.workflowLocale = req.locale || self.defaultLocale;
-      if (!doc.workflowLocale.match(/\-draft$/)) {
+      if (!doc.workflowLocale.match(/-draft$/)) {
         // Always create the draft first, so we can then find it by id successfully
         // via code that is overridden to look for drafts. All the locales get created
         // but we want to return the draft's _id
@@ -153,25 +153,25 @@ module.exports = function(self, options) {
       self.ensureWorkflowLocaleForPathIndex(doc);
     }
   };
-  
+
   // Adjust the slug of a page to take the prefix into account.
   // The UI and/or `pages.newChild` should have done this already, this
   // is a failsafe invoked by `docBeforeSave` and also in tasks.
   //
   // If the document is not a page, or has no locale, nothing happens.
-  
+
   self.ensurePageSlugPrefix = function(doc) {
     var prefix = self.prefixes && self.prefixes[self.liveify(doc.workflowLocale)];
     if (!(prefix && self.apos.pages.isPage(doc) && doc.workflowLocale)) {
       return;
     }
     // Match the first component of the URL
-    matches = doc.slug && doc.slug.match(/^\/([^\/]+)/);
+    var matches = doc.slug && doc.slug.match(/^\/([^/]+)/);
     if (!matches) {
       // No first component or no slug at all
       doc.slug = prefix + (doc.slug || '/' + self.apos.utils.slugify(doc.title));
     } else {
-      existing = matches[1];
+      var existing = matches[1];
       if (('/' + existing) === prefix) {
         // Good to go
       } else {
@@ -200,7 +200,7 @@ module.exports = function(self, options) {
   self.findJoinsInDoc = function(doc) {
     return self.findJoinsInDocSchema(doc).concat(self.findJoinsInAreas(doc));
   };
-  
+
   // Given a doc, invoke `findJoinsInSchema` with that doc and its schema according to
   // its doc type manager, and return the result.
 
@@ -214,7 +214,7 @@ module.exports = function(self, options) {
   // `doc` and `field` properties. `doc` is a reference to the individual widget
   // in question, and `field` is the join field definition for that widget.
   // Only forward joins are returned.
-  
+
   self.findJoinsInAreas = function(doc) {
     var widgets = [];
     self.apos.areas.walk(doc, function(area, dotPath) {
@@ -227,12 +227,12 @@ module.exports = function(self, options) {
         // We already warn about obsolete widgets elsewhere, don't crash
         return;
       }
-      schema = manager.schema;
+      var schema = manager.schema;
       joins = joins.concat(self.findJoinsInSchema(widget, schema));
     });
     return joins;
   };
-  
+
   // Given a doc (or widget) and a schema, find joins described by that schema and
   // return an array in which each element is an object with
   // `doc`, `field` and `value` properties. `doc` is a reference to the doc
@@ -303,20 +303,20 @@ module.exports = function(self, options) {
   self.modulesReady = function(callback) {
     return self.ensureIndexes(callback);
   };
-  
+
   self.ensureIndexes = function(callback) {
 
     return Promise.try(function() {
       return self.apos.docs.db.ensureIndex({ workflowGuid: 1 }, {});
     })
-    .then(workflowGuidWorkflowLocale)
+      .then(workflowGuidWorkflowLocale)
     // depromisify
-    .then(function() {
-      return callback(null);
-    })
-    .catch(function(err) {
-      return callback(err);
-    });
+      .then(function() {
+        return callback(null);
+      })
+      .catch(function(err) {
+        return callback(err);
+      });
 
     // Early versions did not have this index and thus could
     // violate the uniqueness of the guid.locale pair under
@@ -325,13 +325,14 @@ module.exports = function(self, options) {
 
     function workflowGuidWorkflowLocale() {
       return self.apos.docs.db.ensureIndex({ workflowGuid: 1, workflowLocale: 1 }, { sparse: 1, unique: 1 })
-      .catch(function(err) {
-        return resolveDuplicateDocs()
-        .then(function() {
-          // now we can try the index again, recursively
-          return workflowGuidWorkflowLocale();
+        // eslint-disable-next-line handle-callback-err
+        .catch(function(err) {
+          return resolveDuplicateDocs()
+            .then(function() {
+              // now we can try the index again, recursively
+              return workflowGuidWorkflowLocale();
+            });
         });
-      });
     }
 
     function resolveDuplicateDocs() {
@@ -341,133 +342,132 @@ module.exports = function(self, options) {
       return Promise.try(function() {
         return self.apos.locks.lock('apostrophe-workflow:resolveDuplicateDocs');
       })
-      .then(function() {
-        locked = true;
-        return self.apos.docs.db.aggregate([
-          {
-            $match: {
-              workflowGuid: { $exists: 1 }
+        .then(function() {
+          locked = true;
+          return self.apos.docs.db.aggregate([
+            {
+              $match: {
+                workflowGuid: { $exists: 1 }
+              }
             },
-          },
-          {
-            $group: {
-              _id: {
-                workflowGuid: "$workflowGuid",
-                workflowLocale: "$workflowLocale"
-              },
-              slugsAndIds: {
-                $addToSet: { slug: "$slug", _id: "$_id" }
-              },
-              count: {
-                $sum: 1
+            {
+              $group: {
+                _id: {
+                  workflowGuid: "$workflowGuid",
+                  workflowLocale: "$workflowLocale"
+                },
+                slugsAndIds: {
+                  $addToSet: { slug: "$slug", _id: "$_id" }
+                },
+                count: {
+                  $sum: 1
+                }
+              }
+            },
+            {
+              $match: {
+                count: {
+                  $gt: 1
+                }
               }
             }
-          },
-          {
-            $match: {
-              count: {
-                $gt: 1
-              }
-            }
-          }
-        ], { allowDiskUse: true }).toArray();
-      })
-      .then(function(groups) {
-        return Promise.mapSeries(groups, function(group) {
+          ], { allowDiskUse: true }).toArray();
+        })
+        .then(function(groups) {
+          return Promise.mapSeries(groups, function(group) {
           // Shortest slug is the keeper (the one with
           // a digit added is the one created later)
-          group.slugsAndIds.sort(function(a, b) {
-            if (a.slug < b.slug) {
-              return -1;
-            } else if (a.slug > b.slug) {
-              return 1;
-            } else {
-              return 0;
-            }
-          });
-          var id = group.slugsAndIds[0]._id;
-          return Promise.mapSeries(group.slugsAndIds.slice(1), function(slugAndId) {
-            return Promise.try(function() {
-              return self.apos.docs.db.findOne({
-                _id: slugAndId._id
-              });
-            })
-            .then(function(orphan) {
-              if (!orphan) {
-                // Gone already somehow; not our problem
-                return;
+            group.slugsAndIds.sort(function(a, b) {
+              if (a.slug < b.slug) {
+                return -1;
+              } else if (a.slug > b.slug) {
+                return 1;
+              } else {
+                return 0;
               }
-              console.error(orphan._id + ' duplicates the workflowGuid and workflowLocale of\n' + id + ' and has been removed and appended to the\nworkflowGuidAndLocaleDuplicates property of the latter to resolve this race\ncondition.\n\nThis is no longer possible for new inserts thanks to a unique sparse index.\n');
-              idsToNew[orphan._id] = id;
-              return self.apos.docs.db.update({
-                _id: id
-              }, {
-                $push: {
-                  workflowGuidAndLocaleDuplicates: orphan
-                }
-              })
-              .then(function() {
-                return self.apos.docs.db.remove({
-                  _id: orphan._id
+            });
+            var id = group.slugsAndIds[0]._id;
+            return Promise.mapSeries(group.slugsAndIds.slice(1), function(slugAndId) {
+              return Promise.try(function() {
+                return self.apos.docs.db.findOne({
+                  _id: slugAndId._id
                 });
-              });
+              })
+                .then(function(orphan) {
+                  if (!orphan) {
+                    // Gone already somehow; not our problem
+                    return;
+                  }
+                  console.error(orphan._id + ' duplicates the workflowGuid and workflowLocale of\n' + id + ' and has been removed and appended to the\nworkflowGuidAndLocaleDuplicates property of the latter to resolve this race\ncondition.\n\nThis is no longer possible for new inserts thanks to a unique sparse index.\n');
+                  idsToNew[orphan._id] = id;
+                  return self.apos.docs.db.update({
+                    _id: id
+                  }, {
+                    $push: {
+                      workflowGuidAndLocaleDuplicates: orphan
+                    }
+                  })
+                    .then(function() {
+                      return self.apos.docs.db.remove({
+                        _id: orphan._id
+                      });
+                    });
+                });
             });
           });
-        });
-      })
-      .then(function() {
+        })
+        .then(function() {
         // Alas, we still must iterate all docs to look for joins
         // with the orphaned docs and rewrite them, but at least
         // we can limit writes to cases where they are necessary
 
-        var count = 0;
-        return Promise.promisify(function(callback) {
-          return self.apos.migrations.eachDoc({}, function(doc, callback) {
+          return Promise.promisify(function(callback) {
+            return self.apos.migrations.eachDoc({}, function(doc, callback) {
 
-            var modified = false;
+              var modified = false;
 
-            replaceIdsRecursively(doc);
+              replaceIdsRecursively(doc);
 
-            if (!modified) {
-              return setImmediate(callback);
-            }
+              if (!modified) {
+                return setImmediate(callback);
+              }
 
-            return self.apos.docs.db.update({
-              _id: doc._id
-            }, doc, callback);
+              return self.apos.docs.db.update({
+                _id: doc._id
+              }, doc, callback);
 
-            // Recursively replace all occurrences of the ids in this locale
-            // found in the given doc with their new ids per `idsToNew`. This prevents
-            // _id conflicts on insert, even though old data is still in the database
-            // under other locale names
+              // Recursively replace all occurrences of the ids in this locale
+              // found in the given doc with their new ids per `idsToNew`. This prevents
+              // _id conflicts on insert, even though old data is still in the database
+              // under other locale names
 
-            function replaceIdsRecursively(doc) {
-              _.each(doc, function(val, key) {
-                if (key === 'workflowGuidAndLocaleDuplicates') {
+              function replaceIdsRecursively(doc) {
+                _.each(doc, function(val, key) {
+                  if (key === 'workflowGuidAndLocaleDuplicates') {
                   // Do not alter ids inside the historical archive of
                   // the original orphaned docs. -Tom
-                  return;
-                }
-                if ((typeof(val) === 'string') && (val.length < 100)) {
-                  if (idsToNew[val]) {
-                    console.warn('Correcting ' + doc[key] + ' to ' + idsToNew[val] + ' in join');
-                    doc[key] = idsToNew[val];
-                    modified = true;
+                    return;
                   }
-                } else if (val && (typeof(val) === 'object')) {
-                  replaceIdsRecursively(val);
-                }
-              });
-            }
-          }, callback);
-        })();
-      })
-      .finally(function() {
-        if (!locked) {
-          return;
-        }
-        return self.apos.locks.unlock('apostrophe-workflow:resolveDuplicateDocs');
-      });
+                  if ((typeof (val) === 'string') && (val.length < 100)) {
+                    if (idsToNew[val]) {
+                      console.warn('Correcting ' + doc[key] + ' to ' + idsToNew[val] + ' in join');
+                      doc[key] = idsToNew[val];
+                      modified = true;
+                    }
+                  } else if (val && (typeof (val) === 'object')) {
+                    replaceIdsRecursively(val);
+                  }
+                });
+              }
+            }, callback);
+          })();
+        })
+        .finally(function() {
+          if (!locked) {
+            return;
+          }
+          return self.apos.locks.unlock('apostrophe-workflow:resolveDuplicateDocs');
+        });
     }
   };
 
@@ -475,7 +475,7 @@ module.exports = function(self, options) {
   // This is distinct from the versions collection, which becomes more sparse
   // as you move back through time and doesn't always give access to the
   // version that originally preceded a given version.
-  
+
   self.enableCollection = function(callback) {
     self.db = self.apos.db.collection('aposWorkflowCommits');
     var indexes = [
@@ -501,7 +501,7 @@ module.exports = function(self, options) {
   self.getStem = function(dotPath) {
     var stem = dotPath.split(/\./);
     stem.pop();
-    return stem = stem.join('.');
+    return stem.join('.');
   };
 
   // Given a dot path like a.b.5, return
@@ -519,15 +519,15 @@ module.exports = function(self, options) {
   self.guessLocale = function(req) {
     req.locale = self.defaultLocale;
   };
-  
+
   self.enableCrossDomainSessionCache = function() {
     self.crossDomainSessionCache = self.apos.caches.get('apostrophe-workflow-cross-domain-session-cache');
   };
-  
+
   self.acceptCrossDomainSessionToken = function(req) {
     var crossDomainSessionToken = self.apos.launder.string(req.query.workflowCrossDomainSessionToken);
     var sessionData;
-    
+
     return async.series([
       get,
       clear
@@ -543,7 +543,7 @@ module.exports = function(self, options) {
       });
       _.assign(req.session, JSON.parse(sessionData));
       return after();
-      
+
       function after() {
         return req.res.redirect(self.apos.urls.build(req.url, { workflowCrossDomainSessionToken: null }));
       }
@@ -561,7 +561,7 @@ module.exports = function(self, options) {
 
     function clear(callback) {
       return self.crossDomainSessionCache.set(crossDomainSessionToken, null, callback);
-    }    
+    }
   };
-                        
+
 };
