+++ lib/modules/apostrophe-workflow-admin-bar/index.js
@@ -13,7 +13,7 @@ module.exports = {
       if (!result) {
         return result;
       }
-      if (req.locale && req.locale.match(/\-draft$/)) {
+      if (req.locale && req.locale.match(/-draft$/)) {
         return result;
       }
       var notSafeLiveList = [ 'apostrophe-pages', 'apostrophe-tags', 'apostrophe-workflow-manage-modal' ];
@@ -37,4 +37,4 @@ module.exports = {
       return false;
     };
   }
-}
+};
