+++ lib/modules/apostrophe-workflow-areas/index.js
@@ -17,7 +17,7 @@ module.exports = {
               action: 'workflow-force-export-widget'
             }
           ]
-        }        
+        }
       ]);
     };
   }
