+++ lib/modules/apostrophe-workflow-docs/index.js
@@ -29,7 +29,7 @@ module.exports = {
 
     // Solve chicken and egg problem by making sure we have a
     // workflow locale before we test insert permissions
-    
+
     var superTestInsertPermissions = self.testInsertPermissions;
     self.testInsertPermissions = function(req, doc, options) {
       var workflow = self.apos.modules['apostrophe-workflow'];
