+++ lib/modules/apostrophe-workflow-docs/lib/cursor.js
@@ -1,5 +1,3 @@
-var _ = require('lodash');
-
 module.exports = {
   construct: function(self, options) {
     // npm modules should play nice and not use `alias`
@@ -29,6 +27,6 @@ module.exports = {
         self.and({ $or: [ { workflowLocale: setting }, { workflowLocale: { $exists: 0 } } ] });
       }
     });
-        
+
   }
 };
