+++ lib/modules/apostrophe-workflow-groups/index.js
@@ -1,5 +1,4 @@
 var _ = require('lodash');
-var async = require('async');
 
 module.exports = {
 
@@ -13,7 +12,7 @@ module.exports = {
       superModulesReady();
       self.workflowModifyPermissionsField();
     };
-    
+
     self.workflowModifyPermissionsField = function() {
       var workflow = self.apos.modules['apostrophe-workflow'];
       var permissions = _.find(self.schema, { name: 'permissions' });
