+++ lib/modules/apostrophe-workflow-pages/index.js
@@ -27,7 +27,7 @@ module.exports = {
       }
       var locale = workflow.liveify(req.locale);
       if (_.has(workflow.prefixes, locale)) {
-        var matches = slug.match(/^(\/[^\/]+)(\/?)$/);
+        var matches = slug.match(/^(\/[^/]+)(\/?)$/);
         if (matches && (workflow.prefixes[locale] === matches[1])) {
           // Something like /en/, leave it alone,
           // it's a localized homepage. However if the
@@ -87,15 +87,15 @@ module.exports = {
             action: 'workflow-commit'
           }
         ].concat(
-          (workflow.localized && (verb === 'edit')) ?
-          [
-            {
-              label: 'History',
-              action: 'workflow-history'
-            }
-          ] : []
-        ).concat(workflow.localized ?
-          [
+          (workflow.localized && (verb === 'edit'))
+            ? [
+              {
+                label: 'History',
+                action: 'workflow-history'
+              }
+            ] : []
+        ).concat(workflow.localized
+          ? [
             {
               label: 'Force Export',
               action: 'workflow-force-export'
@@ -104,13 +104,13 @@ module.exports = {
         )
       });
       return controls;
-    }    
+    }
 
     var superMovePermissions = self.movePermissions;
 
     // "Based on `req`, `moved`, `data.oldParent` and `data.parent`, decide whether
     // this move should be permitted. If it should not be, report an error." The `apostrophe-pages`
-    // module did this already, but we must consider the impact on all locales. 
+    // module did this already, but we must consider the impact on all locales.
 
     self.movePermissions = function(req, moved, data, options, callback) {
       return superMovePermissions(req, moved, data, options, function(err) {
@@ -164,7 +164,7 @@ module.exports = {
 
     // On the initial invocation of `apos.pages.move`, modify the criteria and filters
     // to ensure only the relevant locale is in play
-    
+
     var superBeforeMove = self.beforeMove;
     self.beforeMove = function(req, moved, target, position, options, callback) {
       return superBeforeMove(req, moved, target, position, options, function(err) {
@@ -195,7 +195,6 @@ module.exports = {
         if (info.options.workflowRecursing) {
           return callback(null);
         }
-        var ids = _.pluck(info.changed || [], '_id');
         if (!moved.workflowGuid) {
           return callback(null);
         }
@@ -226,7 +225,7 @@ module.exports = {
           });
         }
         function invoke(callback) {
-          // Reinvoke apos.pages.move 
+          // Reinvoke apos.pages.move
           return async.eachSeries(_.keys(locales), function(locale, callback) {
             var _options = _.clone(info.options);
             _options.criteria = _.assign({}, info.options.criteria || {}, { workflowLocale: locale });
@@ -237,7 +236,7 @@ module.exports = {
         }
       });
     };
-  
+
     // `implementParkAll` must be reimplemented to cover all of
     // the locales and to adjust the slugs if the prefixes option
     // is in use
@@ -251,7 +250,7 @@ module.exports = {
       }
       var workflow = self.apos.modules['apostrophe-workflow'];
       return async.eachSeries(_.keys(workflow.locales), function(locale, callback) {
-        if (locale.match(/\-draft/)) {
+        if (locale.match(/-draft/)) {
           // Skip the draft locales, the doc will automatically get created
           // in both draft and live locales
           return setImmediate(callback);
@@ -291,7 +290,7 @@ module.exports = {
         }
       }, callback);
     };
-    
+
     var superGetBaseUrl = self.getBaseUrl;
     // Return the appropriate base URL for constructing absolute
     // URLs in the relevant locale, if the `hostnames` option is
@@ -307,7 +306,7 @@ module.exports = {
       // If the request URL was on a particular port, the
       // new URL should be on that port too. Helpful for
       // debugging locally
-      var matches = (req.get('host') || '').match(/\:(\d+$)/);
+      var matches = (req.get('host') || '').match(/:(\d+$)/);
       var port;
       if (matches) {
         port = matches[1];
