+++ lib/modules/apostrophe-workflow-pages/public/js/editor.js
@@ -9,7 +9,7 @@ apos.define('apostrophe-pages-editor', {
         return self.save(function(err) {
           if (err) {
             self.submitting = false;
-            return;
+
           }
           // Never reached due to redirect
         });
@@ -20,7 +20,7 @@ apos.define('apostrophe-pages-editor', {
         return self.save(function(err) {
           if (err) {
             self.committing = false;
-            return;
+
           }
           // Never reached due to redirect
         });
@@ -30,7 +30,7 @@ apos.define('apostrophe-pages-editor', {
         return self.save(function(err) {
           if (err) {
             self.forceExporting = false;
-            return;
+
           }
           // Never reached due to redirect
         });
@@ -57,7 +57,6 @@ apos.define('apostrophe-pages-editor', {
       });
     };
 
-    var superAfterSave = self.afterSave;
     self.afterSave = function(callback) {
       if (self.submitting) {
         return apos.modules['apostrophe-workflow'].submit([ self.savedPage._id ], callback);
