+++ lib/modules/apostrophe-workflow-permissions/index.js
@@ -1,5 +1,4 @@
 var _ = require('lodash');
-var async = require('async');
 
 module.exports = {
 
@@ -35,14 +34,14 @@ module.exports = {
       // any access to a private locale
       var locale = workflow.locales[req.locale || workflow.defaultLocale];
       if (
-        (!locale) || 
-        (locale.private && 
+        (!locale) ||
+        (locale.private &&
           ((!req.user) || (
             (!req.user._permissions['private-locales']) &&
             (!req.user._permissions['admin'])
           ))
-         )
-        ) {
+        )
+      ) {
         info.response = info._false;
         return;
       }
@@ -65,7 +64,7 @@ module.exports = {
       // in draft vs. live locales
       if (verb === 'publish') {
         verb = 'edit';
-        action = action.replace(/^publish\-/, 'edit-');
+        action = action.replace(/^publish-/, 'edit-');
       }
       if (!_.contains(workflow.includeVerbs, verb)) {
         return;
@@ -88,18 +87,18 @@ module.exports = {
       // publish is not a separate verb in workflow since we already control whether you can edit
       // in draft vs. live locales
       if (info.verb === 'publish') {
-        action = action.replace(/^publish\-/, 'edit-');
+        action = action.replace(/^publish-/, 'edit-');
       }
 
       if (!(req.user && req.user._permissionsLocales)) {
         info.response = info._false;
         return;
       }
-      
+
       var adminAction = 'admin-' + info.type;
 
       // 'VERB', 'VERB-this-type' or 'admin-this-type' is acceptable
-      var permissionsLocales = _.assign({}, 
+      var permissionsLocales = _.assign({},
         req.user._permissionsLocales[action] || {},
         req.user._permissionsLocales[adminAction] || {},
         req.user._permissionsLocales[verb] || {}
@@ -113,7 +112,7 @@ module.exports = {
       if (event === 'criteria') {
         info.response = { $and: [ info.response, { workflowLocale: { $in: _.keys(permissionsLocales) } } ] };
       } else {
-        var object = info.object || info.newObject;
+        object = info.object || info.newObject;
         if (object) {
           if (!permissionsLocales[object.workflowLocale]) {
             info.response = info._false;
@@ -124,6 +123,6 @@ module.exports = {
       }
 
     };
-    
+
   }
 };
