+++ lib/modules/apostrophe-workflow-pieces/index.js
@@ -1,12 +1,11 @@
 var _ = require('lodash');
-var async = require('async');
 
 module.exports = {
 
   improve: 'apostrophe-pieces',
 
   canEditTrash: true,
-  
+
   beforeConstruct: function(self, options) {
 
     function onlyIf(type) {
@@ -39,9 +38,9 @@ module.exports = {
     ].concat(options.addBatchOperations || []);
 
   },
-  
+
   construct: function(self, options) {
-    
+
     var superGetEditControls = self.getEditControls;
     self.getEditControls = function(req) {
       return upgradeControls(req, superGetEditControls(req), 'edit');
@@ -62,7 +61,7 @@ module.exports = {
       if (save) {
         save.label = 'Save Draft';
       }
-      
+
       controls.push({
         type: 'dropdown',
         label: 'Workflow',
@@ -78,16 +77,16 @@ module.exports = {
           {
             label: 'Commit',
             action: 'workflow-commit'
-          },
-        ].concat((workflow.localized && (verb === 'edit')) ?
-          [
+          }
+        ].concat((workflow.localized && (verb === 'edit'))
+          ? [
             {
               label: 'History',
               action: 'workflow-history'
             }
           ] : []
-        ).concat(workflow.localized ?
-          [
+        ).concat(workflow.localized
+          ? [
             {
               label: 'Force Export',
               action: 'workflow-force-export'
