+++ lib/modules/apostrophe-workflow-pieces/public/js/editor-modal.js
@@ -29,19 +29,22 @@ apos.define('apostrophe-pieces-editor-modal', {
       self.workflowControlsVisibility();
       return superBeforeShow(callback);
     };
-    
+
     // Save the modal normally, then invoke the given callback.
     // The callback is invoked only if the modal is saved successfully
 
     self.workflowSaveThen = function(callback) {
       self.workflowBeforeDisplayResponse = callback;
       return self.save(function(err) {
+        if (err) {
+          console.error(err);
+        }
         self.workflowBeforeDisplayResponse = null;
       });
     };
 
     var superDisplayResponse = self.displayResponse;
-    
+
     // Invoke `workflowBeforeDisplayResponse` if present, then
     // respond to the saving of the piece in the normal way
     self.displayResponse = function(result, callback) {
@@ -56,7 +59,7 @@ apos.define('apostrophe-pieces-editor-modal', {
         return superDisplayResponse(result, callback);
       });
     };
-    
+
     self.workflowControlsVisibility = function() {
       var workflow = apos.modules['apostrophe-workflow'];
       return workflow.api('committable', { type: self.name, id: self._id }, function(results) {
