+++ lib/modules/apostrophe-workflow-schemas/index.js
@@ -1,5 +1,4 @@
 var _ = require('lodash');
-var async = require('async');
 
 module.exports = {
 
@@ -11,7 +10,6 @@ module.exports = {
 
   construct: function(self, options) {
     self.workflowAddPermissionsFieldType = function() {
-      var workflow = self.apos.modules['apostrophe-workflow'];
       self.apos.schemas.addFieldType({
         name: 'apostrophe-workflow-permissions',
         partial: self.workflowPermissionsPartial,
@@ -46,7 +44,7 @@ module.exports = {
           }
         }
         return true;
-      })
+      });
       return self.partial('workflow-permissions-schema-field', data);
     };
 
@@ -64,10 +62,10 @@ module.exports = {
         object[name] = _.filter(data[name], function(choice) {
           return _.contains(_.pluck(field.choices, 'value'), choice);
         });
-        
+
         var permissionsLocales = {};
         var raw = data[name + 'Locales'];
-        if ((!raw) || (typeof(raw) !== 'object')) {
+        if ((!raw) || (typeof (raw) !== 'object')) {
           return setImmediate(callback);
         }
         _.each(raw, function(locales, permission) {
