+++ lib/modules/apostrophe-workflow-schemas/public/js/user.js
@@ -16,110 +16,111 @@ apos.define('apostrophe-schemas', {
       return apos.modules['apostrophe-workflow'].api('get-live',
         _.assign(_.pick(object, 'workflowGuid', 'workflowLocale'), { resolveRelationshipsToDraft: true }),
         function(data) {
-        if (data.status !== 'ok') {
-          return fail();
-        }
-        // Give each field in the schema a read-only twin that is populated with
-        // the live version of the same data, and provide UI for peeking at it to
-        // see what's different.
-        //
-        // This is actually simpler than generating a crappy plaintext representation
-        // of the same content. -Tom
-        
-        return async.eachSeries(schema, function(field, callback) {
-          var $draft = self.findFieldset($el, field.name);
-          // The visible live version for preview purposes
-          var $live = $draft.clone();
-          // A pristine copy for revert purposes
-          var $livePristine = $draft.clone();
-          // Nest the twin in an apos-field so that findFieldset for the original schema doesn't mistakenly descend into it
-          var $wrapper = $('<div class="apos-field apos-workflow-live-field" data-apos-workflow-live-field data-name="workflow-live-' + field.name + '"></div>');
-          $wrapper.append($live);
-          $wrapper.append($('<div class="apos-workflow-live-field-mask"></div>'));
-          $draft.after($wrapper);
-          var draftValue = object[field.name];
-          var liveValue = data.doc[field.name];
-          if (_.isEqual(draftValue, liveValue)) {
-            // No modification
-          } else if (!(draftValue || liveValue)) {
-            // Simple false positives like undefined vs. false
-          } else if (Array.isArray(draftValue) && (!draftValue.length) && (!liveValue)) {
-            // undefined vs. empty array
-          } else {
-            // OK, it's a modification
-            $draft.addClass('apos-workflow-field-changed');
+          if (data.status !== 'ok') {
+            return fail();
           }
+          // Give each field in the schema a read-only twin that is populated with
+          // the live version of the same data, and provide UI for peeking at it to
+          // see what's different.
+          //
+          // This is actually simpler than generating a crappy plaintext representation
+          // of the same content. -Tom
 
-          var $draftControls = $(
-            '<span class="apos-workflow-field-controls">' +
+          return async.eachSeries(schema, function(field, callback) {
+            var $draft = self.findFieldset($el, field.name);
+            // The visible live version for preview purposes
+            var $live = $draft.clone();
+            // A pristine copy for revert purposes
+            var $livePristine = $draft.clone();
+            // Nest the twin in an apos-field so that findFieldset for the original schema doesn't mistakenly descend into it
+            var $wrapper = $('<div class="apos-field apos-workflow-live-field" data-apos-workflow-live-field data-name="workflow-live-' + field.name + '"></div>');
+            $wrapper.append($live);
+            $wrapper.append($('<div class="apos-workflow-live-field-mask"></div>'));
+            $draft.after($wrapper);
+            var draftValue = object[field.name];
+            var liveValue = data.doc[field.name];
+            if (_.isEqual(draftValue, liveValue)) {
+            // No modification
+            } else if (!(draftValue || liveValue)) {
+            // Simple false positives like undefined vs. false
+            } else if (Array.isArray(draftValue) && (!draftValue.length) && (!liveValue)) {
+            // undefined vs. empty array
+            } else {
+            // OK, it's a modification
+              $draft.addClass('apos-workflow-field-changed');
+            }
+
+            var $draftControls = $(
+              '<span class="apos-workflow-field-controls">' +
               '<select class="apos-workflow-field-state-control" data-apos-workflow-field-state-control>' +
                 '<option value="draft" selected>Draft</option>' +
                 '<option value="live">Live</option>' +
               '</select>' +
             '</span>'
-          );
+            );
 
-          $draft.find('label').prepend($draftControls);
+            $draft.find('label').prepend($draftControls);
 
-          var $liveControls = $(
-            '<span class="apos-workflow-field-controls">' +
+            var $liveControls = $(
+              '<span class="apos-workflow-field-controls">' +
               '<select class="apos-workflow-field-state-control" data-apos-workflow-field-state-control>' +
                 '<option value="draft">Draft</option>' +
                 '<option value="live" selected>Live</option>' +
               '</select>' +
               '<a href="#" class="apos-workflow-revert-field" data-apos-workflow-revert></a>' +
             '</span>'
-          );
+            );
 
-          $live.find('label').prepend($liveControls);
-          
-          $draftControls.find('[data-apos-workflow-field-state-control]').on('change', function () {
-            var $select = $(this);
-            var state = $select.val();
-            if (state === 'live') {
-              $live.show();
-              $draft.hide();
-              // So when we toggle back to this we see the right initial state again
-              $select.val('draft');
-            }
-          });
+            $live.find('label').prepend($liveControls);
 
-          $liveControls.find('[data-apos-workflow-field-state-control]').on('change', function () {
-            var $select = $(this);
-            var state = $select.val();
-            if (state === 'draft') {
+            $draftControls.find('[data-apos-workflow-field-state-control]').on('change', function () {
+              var $select = $(this);
+              var state = $select.val();
+              if (state === 'live') {
+                $live.show();
+                $draft.hide();
+                // So when we toggle back to this we see the right initial state again
+                $select.val('draft');
+              }
+            });
+
+            $liveControls.find('[data-apos-workflow-field-state-control]').on('change', function () {
+              var $select = $(this);
+              var state = $select.val();
+              if (state === 'draft') {
+                $live.hide();
+                $draft.show();
+                // So when we toggle back to this we see the right initial state again
+                $select.val('live');
+              }
+            });
+
+            $liveControls.on('click', '[data-apos-workflow-revert]:first', function() {
+              var $revert = $livePristine.clone();
+              $draft.replaceWith($revert);
               $live.hide();
-              $draft.show();
-              // So when we toggle back to this we see the right initial state again
-              $select.val('live');
-            }
-          });
+              apos.schemas.populate($el, [ field ], data.doc, callback);
+              return false;
+            });
 
-          $liveControls.on('click', '[data-apos-workflow-revert]:first', function() {
-            var $revert = $livePristine.clone();
-            $draft.replaceWith($revert);
-            $live.hide();
-            apos.schemas.populate($el, [ field ], data.doc, callback);
-            return false;
-          });
+            // Populate it via a schema of just that one field, so we don't reinvent any wheels but it still
+            // fits into its separate place in the DOM. Use a hint property to avoid infinite recursion on this.
+            // We use hints because the blessing mechanism on the server side ignores them.
 
-          // Populate it via a schema of just that one field, so we don't reinvent any wheels but it still
-          // fits into its separate place in the DOM. Use a hint property to avoid infinite recursion on this.
-          // We use hints because the blessing mechanism on the server side ignores them.
+            return apos.schemas.populate($wrapper, [ _.merge({ hints: { workflowRecursing: true } }, field) ], data.doc, callback);
 
-          return apos.schemas.populate($wrapper, [ _.merge({ hints: { workflowRecursing: true } }, field) ], data.doc, callback);
-
-        }, function(err) {
-          if (err) {
+          }, function(err) {
+            if (err) {
             // Could be an inconsistent state, so treat it as a proper error
-            apos.notify('An error occurred. Please try again.', { type: 'error' });
-            return callback(err);
-          }
-          return next();
+              apos.notify('An error occurred. Please try again.', { type: 'error' });
+              return callback(err);
+            }
+            return next();
+          });
+          // eslint-disable-next-line handle-callback-err
+        }, function(err) {
+          return fail();
         });
-      }, function(err) {
-        return fail();
-      });
       function fail() {
         console.error('Nonfatal error: cannot fetch live version of draft doc.');
         return next();
@@ -135,14 +136,14 @@ apos.define('apostrophe-schemas', {
     self.afterPopulate = function($el, schema, object, callback) {
       // Make sure people can't edit the preview of the live version of the field
       $el.find('[data-apos-workflow-live-field] > fieldset *').each(function() {
-        if (typeof this.disabled != "undefined" ) {
+        if (typeof this.disabled !== "undefined") {
           this.disabled = true;
         }
       });
       // But allow the toggles to be clicked
       // TOM PLEASE MAKE THIS WORK
-      $el.find('[data-apos-workflow-live-field] [data-apos-workflow-field-state-control], [data-apos-workflow-live-field] [data-apos-workflow-field-state-control] *').each(function(){
-        $(this).prop( "disabled", false );
+      $el.find('[data-apos-workflow-live-field] [data-apos-workflow-field-state-control], [data-apos-workflow-live-field] [data-apos-workflow-field-state-control] *').each(function() {
+        $(this).prop("disabled", false);
       });
 
       // Try to give away the focus if it arrives via keyboard on an anchor element
@@ -176,7 +177,6 @@ apos.define('apostrophe-schemas', {
       var currentSlug = $slug.val();
       var components = currentSlug.split('/');
       var currentSlugTitle = components.pop();
-      var locale = workflow.locale;
       var prefix = workflow.options.prefixes[liveLocale];
       $title.on('change keyup paste', function(e) {
         var slug = $slug.val();
@@ -185,16 +185,16 @@ apos.define('apostrophe-schemas', {
         // Original mechanism to sync last component of slug with title if
         // and only if this was the case before editing began
         if ((originalTitle === '') || (currentSlugTitle === apos.utils.slugify(originalTitle))) {
-          slug = slug.replace(/[^\/]*$/, apos.utils.slugify($title.val()));
+          slug = slug.replace(/[^/]*$/, apos.utils.slugify($title.val()));
         }
-        
+
         // Match the first component of the URL
-        matches = slug && slug.match(/^\/([^\/]+)/);
+        matches = slug && slug.match(/^\/([^/]+)/);
         if (!matches) {
           // No first component or no slug at all
-          slug = '/' + doc.workflowLocale + (slug || '/' + self.apos.utils.slugify(title));
+          slug = prefix + (slug || '/' + self.apos.utils.slugify(title));
         } else {
-          existing = matches[1];
+          var existing = matches[1];
           if (('/' + existing) === workflow.options.prefixes[liveLocale]) {
             // Great
           } else {
@@ -215,11 +215,11 @@ apos.define('apostrophe-schemas', {
         convert: self.workflowPermissionsConvert
       });
     };
-    
+
     self.findWorkflowPermissionsCheckbox = function($fieldset, name, val) {
       return $fieldset.find('input[name="' + name + '"][value="' + val + '"]');
     };
-    
+
     self.findWorkflowPermissionsLocaleSelect = function($fieldset, name, val, locale) {
       return $fieldset.find('select[name="' + name + 'Locales[' + val + '][' + locale + ']"]');
     };
@@ -231,8 +231,8 @@ apos.define('apostrophe-schemas', {
       });
       _.each(object[name + 'Locales'] || {}, function(locales, permission) {
         _.each(locales, function(locale, localeName) {
-          var $select = self.findWorkflowPermissionsLocaleSelect($fieldset, name, permission, localeName.replace(/\-draft$/, ''));
-          var isDraft = localeName.match(/\-draft/);
+          var $select = self.findWorkflowPermissionsLocaleSelect($fieldset, name, permission, localeName.replace(/-draft$/, ''));
+          var isDraft = localeName.match(/-draft/);
           if (isDraft && ($select.val() !== 'commit')) {
             $select.val('edit');
           } else if (!isDraft) {
@@ -269,7 +269,7 @@ apos.define('apostrophe-schemas', {
             data[name + 'Locales'][choice.value] = {};
           }
           _.each(field.locales, function(locale, localeName) {
-            if (localeName.match(/\-draft$/)) {
+            if (localeName.match(/-draft$/)) {
               return;
             }
             var $select = self.findWorkflowPermissionsLocaleSelect($fieldset, name, choice.value, localeName);
@@ -284,7 +284,7 @@ apos.define('apostrophe-schemas', {
         }
       });
       return setImmediate(callback);
-    };      
+    };
 
   }
 });
