--- lib/permissions-schema-field.js
@@ -1,8 +1,0 @@
-var _ = require('lodash');
-var async = require('async');
-
-module.exports = function(self, options) {
-
-
-
-}
