+++ lib/removeDotPathViaSplice.js
@@ -29,4 +29,3 @@ module.exports = function(o, dotPath) {
     return false;
   }
 };
-
