+++ lib/routes.js
@@ -12,7 +12,7 @@ module.exports = function(self, options) {
       return res.send({ status: 'ok', commitId: commitId, title: title });
     });
   });
-  
+
   self.route('post', 'batch-commit', function(req, res) {
     return self.apos.modules['apostrophe-jobs'].run(req, function(req, id, callback) {
       return self.commitLatest(req, id, callback);
@@ -22,13 +22,13 @@ module.exports = function(self, options) {
       }
     });
   });
-  
+
   self.route('post', 'export', function(req, res) {
     var id = self.apos.launder.id(req.body.id);
     return self.export(req, id, req.body.locales, function(err, results) {
       if (err) {
         return res.send({
-          status: (typeof(err) === 'string') ? err : 'error'
+          status: (typeof (err) === 'string') ? err : 'error'
         });
       }
       return res.send(_.assign({
@@ -36,7 +36,7 @@ module.exports = function(self, options) {
       }, results));
     });
   });
-  
+
   self.route('post', 'batch-export', function(req, res) {
     return self.apos.modules['apostrophe-jobs'].run(req, function(req, id, callback) {
       return self.export(req, id, req.body.locales, callback);
@@ -46,7 +46,7 @@ module.exports = function(self, options) {
       }
     });
   });
-  
+
   self.route('post', 'force-export', function(req, res) {
     return self.forceExport(req, req.body.id, req.body.locales, function(err, results) {
       if (err) {
@@ -71,7 +71,7 @@ module.exports = function(self, options) {
       }
     });
   });
-    
+
   self.route('post', 'force-export-widget', function(req, res) {
     if (!req.user) {
       // Confusion to the enemy
@@ -82,10 +82,10 @@ module.exports = function(self, options) {
     var locales = [];
     var success = [];
     var errors = [];
-    var drafts, original;
+    var original;
     if (Array.isArray(req.body.locales)) {
       locales = _.filter(req.body.locales, function(locale) {
-        return ((typeof(locale) === 'string') && (_.has(self.locales, locale)));
+        return ((typeof (locale) === 'string') && (_.has(self.locales, locale)));
       });
     }
     locales = _.map(locales, function(locale) {
@@ -127,7 +127,7 @@ module.exports = function(self, options) {
       return async.eachSeries(locales, function(locale, callback) {
 
         var resolvedOriginal, draft;
-        
+
         // Our own modifiable copy to safely pass to `resolveToDestination`
         resolvedOriginal = _.cloneDeep(original);
 
@@ -168,7 +168,7 @@ module.exports = function(self, options) {
 
           var parentDotPath = getParentDotPath(widgetInfo.dotPath);
           if (deep(draft, parentDotPath)) {
-            deep(draft, widgetInfo.dotPath, widgetInfo.object); 
+            deep(draft, widgetInfo.dotPath, widgetInfo.object);
             success.push(self.liveify(locale));
           } else if (addMissingArea()) {
             // Great
@@ -178,7 +178,7 @@ module.exports = function(self, options) {
           }
 
           return callback(null);
-          
+
           function addMissingArea() {
             // currently parentDotPath ends in .items
             var areaDotPath = getParentDotPath(parentDotPath);
@@ -195,15 +195,15 @@ module.exports = function(self, options) {
               return false;
             }
           }
-          
+
           function getParentDotPath(dotPath) {
-            return dotPath.replace(/^[^\.]+$|\.[^\.]+$/, '');
+            return dotPath.replace(/^[^.]+$|\.[^.]+$/, '');
           }
-          
+
         }
 
         function update(callback) {
-          return self.apos.docs.update(req, draft, callback); 
+          return self.apos.docs.update(req, draft, callback);
         }
 
       }, callback);
@@ -235,21 +235,21 @@ module.exports = function(self, options) {
       }
       return res.send({ status: 'ok', doc: live });
     });
-    
+
     function get(callback) {
       return self.apos.docs.find(req, { workflowGuid: guid, workflowLocale: locale }).trash(null).published(null).workflowLocale(locale).toObject(function(err, _live) {
         live = _live;
         return callback(err);
       });
     }
-    
+
     function ids(callback) {
       if (!req.body.resolveRelationshipsToDraft) {
         return callback(null);
       }
       return self.resolveRelationships(req, live, self.draftify(locale), callback);
     }
-    
+
     function fail(err) {
       console.error(err);
       return res.send({ status: 'error' });
@@ -292,9 +292,9 @@ module.exports = function(self, options) {
         }
         return res.send({ status: 'ok', url: doc._url });
       }
-    );
+      );
   });
-      
+
   self.route('post', 'submit', function(req, res) {
     if (!req.user) {
       // Confusion to the enemy
@@ -373,6 +373,10 @@ module.exports = function(self, options) {
       return res.status(404).send('not found');
     }
     return self.getSubmitted(req, {}, function(err, submitted) {
+      if (err) {
+        console.error(err);
+        return;
+      }
       return res.send(self.render(req, 'manage-modal.html', { submitted: submitted, label: self.locales[req.locale].label || req.locale }));
     });
   });
@@ -397,6 +401,7 @@ module.exports = function(self, options) {
       // Confusion to the enemy
       return res.status(404).send('not found');
     }
+    var localizations;
     var workflowGuid = self.apos.launder.id(req.body.workflowGuid);
     var crossDomainSessionToken;
     return async.series([ getLocalizations, getCrossDomainSessionToken ], function(err) {
@@ -405,7 +410,7 @@ module.exports = function(self, options) {
         res.status(500).send('error');
       }
       return res.send(self.render(req, 'locale-picker-modal.html', { localizations: localizations, nestedLocales: self.nestedLocales, crossDomainSessionToken: crossDomainSessionToken }));
-    })
+    });
     function getLocalizations(callback) {
       return self.getLocalizations(req, workflowGuid, true, function(err, _localizations) {
         localizations = _localizations;
@@ -414,7 +419,6 @@ module.exports = function(self, options) {
     }
     function getCrossDomainSessionToken(callback) {
       crossDomainSessionToken = self.apos.utils.generateId();
-      var sessionData = req.session;
       return self.crossDomainSessionCache.set(crossDomainSessionToken, JSON.stringify(req.session), 60 * 60, callback);
     }
   });
@@ -428,7 +432,7 @@ module.exports = function(self, options) {
     var index = self.apos.launder.integer(req.body.index);
     var total = self.apos.launder.integer(req.body.total);
     var lead = self.apos.launder.boolean(req.body.lead);
-    var draft, live, modifiedFields, related;
+    var draft, live, modifiedFields;
     return async.series([
       getDraftAndLive,
       getModifiedFields
@@ -452,7 +456,7 @@ module.exports = function(self, options) {
         preview: preview
       }));
     });
-    
+
     function getDraftAndLive(callback) {
       // We get both the same way the commit route does, for the sake of the permissions check,
       // so it doesn't initially appear that someone can be sneaky (although they can't really)
@@ -462,7 +466,7 @@ module.exports = function(self, options) {
         return callback(err);
       });
     }
-                
+
     function getModifiedFields(callback) {
       return self.getModifiedFields(req, draft, live, function(err, _modifiedFields) {
         modifiedFields = _modifiedFields;
@@ -470,7 +474,7 @@ module.exports = function(self, options) {
       });
     }
   });
-  
+
   // Given doc ids in req.body.ids, send back an object with
   // array properties, `modified` and `unmodified`, containing the
   // editable, modified draft doc ids and the editable, unmodified
@@ -491,7 +495,7 @@ module.exports = function(self, options) {
   //
   // For convenience, the ids sent to this method can be either
   // live or draft.
-  
+
   self.route('post', 'editable', function(req, res) {
     if (!req.user) {
       return res.status(404).send('notfound');
@@ -552,7 +556,7 @@ module.exports = function(self, options) {
       return res.send({ status: 'ok' });
     });
   });
-  
+
   // Given body.id and body.exportLocales, this route answers with an object
   // with an ids property containing an array of related ids that have
   // never been exported to at least one of the given locales from this locale.
@@ -565,9 +569,8 @@ module.exports = function(self, options) {
 
   self.route('post', 'related-unexported', function(req, res) {
     var id = self.apos.launder.id(req.body.id);
-    
+
     var exportLocales = self.apos.launder.strings(req.body.exportLocales);
-    var commit;
     var doc;
     var docIds = [];
     var ids = [];
@@ -586,10 +589,9 @@ module.exports = function(self, options) {
       }
       return res.send({ status: 'ok', ids: ids });
     });
-    
+
     function getDoc(callback) {
       return self.findDocAndCommit(req, id, function(err, _doc, _commit) {
-        commit = _commit;
         doc = _doc;
         if (!doc) {
           return callback('notfound');
@@ -615,7 +617,7 @@ module.exports = function(self, options) {
         return callback(err);
       });
     }
-    
+
     // Due to join projections we often do not know the guids
     // of the related docs yet, and even sometimes the type as well
     function getWorkflowGuids(callback) {
@@ -639,12 +641,12 @@ module.exports = function(self, options) {
         related = _.filter(related, function(relatedOne) {
           return relatedOne.type && relatedOne.workflowGuid && self.includeType(relatedOne.type);
         });
-        return callback(null); 
+        return callback(null);
       });
     }
-    
+
     function getLocalizations(callback) {
-      return self.apos.docs.db.find({ 
+      return self.apos.docs.db.find({
         workflowGuid: { $in: _.pluck(related, 'workflowGuid') }
       }).toArray(function(err, localized) {
         if (err) {
@@ -661,10 +663,9 @@ module.exports = function(self, options) {
           }
           // We chose to defer this to here rather than
           // making the query picky with a regex
-          if (!localization.workflowLocale.match(/\-draft$/)) {
+          if (!localization.workflowLocale.match(/-draft$/)) {
             return;
           }
-          var live = self.liveify(localization.workflowLocale);
           if (!localesAccountedFor[localization.workflowGuid]) {
             localesAccountedFor[localization.workflowGuid] = 0;
           }
@@ -679,7 +680,7 @@ module.exports = function(self, options) {
           }
           if (!localization.trash) {
             localesAccountedFor[localization.workflowGuid]++;
-            return;
+
           }
         });
         _.each(related, function(relatedOne) {
@@ -690,7 +691,7 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function getLatestCommits(callback) {
       return async.eachSeries(docIds, function(docId, callback) {
         return self.db.find({ 'from._id': docId }, { _id: 1 }).sort({ createdAt: -1 }).toArray(function(err, commits) {
@@ -704,18 +705,18 @@ module.exports = function(self, options) {
         });
       }, callback);
     }
-    
+
     function fail(err) {
       console.error(err);
       return res.send({ status: 'error' });
     }
   });
- 
+
   // Given a doc `id`, this route replies with
   // `{ locales: [ 'en', 'fr' ] }`, where the returned
   // locales are those for which the current user has
   // editing privileges for docs of the same type as `id`.
-  
+
   self.route('post', 'editable-locales', function(req, res) {
     return self.apos.docs.find(req, { _id: req.body.id }).toObject(function(err, doc) {
       if (err) {
@@ -730,7 +731,7 @@ module.exports = function(self, options) {
         locales: self.getEditableLocales(req, doc)
       });
     });
-  });  
+  });
 
   self.route('post', 'review-modal', function(req, res) {
 
@@ -740,7 +741,6 @@ module.exports = function(self, options) {
     }
 
     var id = self.apos.launder.id(req.body.id);
-    var doc;
     var commit;
     var modifiedFields;
 
@@ -759,7 +759,7 @@ module.exports = function(self, options) {
       }
       return res.send(self.render(req, 'review-modal.html', { preview: preview, commit: commit, doc: commit.to, modifiedFields: modifiedFields }));
     });
-    
+
     function find(callback) {
       return self.findDocAndCommit(req, id, function(err, doc, _commit) {
         if (err) {
@@ -772,7 +772,7 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function getModifiedFields(callback) {
       return self.getModifiedFields(req, commit.to, commit.from, function(err, _modifiedFields) {
         if (err) {
@@ -782,7 +782,7 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function after(callback) {
       return self.after(req, [ commit.to ], callback);
     }
@@ -808,7 +808,7 @@ module.exports = function(self, options) {
       return res.send(self.render(req, 'export-modal.html', { commit: commit, nestedLocales: nestedLocales }));
     });
   });
-  
+
   self.route('post', 'batch-export-modal', function(req, res) {
     if (!req.user) {
       // Confusion to the enemy
@@ -862,7 +862,7 @@ module.exports = function(self, options) {
     }
     // doc id
     var id = self.apos.launder.id(req.body.id);
-    
+
     return self.findDocs(req, { _id: id }).toObject(function(err, doc) {
       if (err) {
         console.error(err);
@@ -898,7 +898,7 @@ module.exports = function(self, options) {
       return res.send(self.render(req, 'batch-force-export-modal.html', { ids: ids, locale: self.liveify(doc.workflowLocale), nestedLocales: nestedLocales }));
     });
   });
-  
+
   self.route('post', 'diff', function(req, res) {
 
     if (!req.user) {
@@ -924,15 +924,15 @@ module.exports = function(self, options) {
         console.error(err);
         return res.send({ status: 'error' });
       }
-      
+
       return res.send({
         status: 'ok',
         diff: diff,
         id: id
       });
-      
+
     });
-    
+
     function getContent(callback) {
       if (commitId) {
         return getCommit(callback);
@@ -940,7 +940,7 @@ module.exports = function(self, options) {
         return getDraftAndLive(callback);
       }
     }
-    
+
     function getCommit(callback) {
       return self.findDocAndCommit(req, commitId, function(err, doc, commit) {
         if (err) {
@@ -966,15 +966,15 @@ module.exports = function(self, options) {
         return callback(null);
       });
     }
-    
+
     function resolveRelationships(callback) {
       var resolve = [];
       // You'd think it would be obvious which of these has a -draft locale suffix,
       // but let's be flexible to allow for different scenarios
-      if (!draft.workflowLocale.match(/\-draft$/)) {
+      if (!draft.workflowLocale.match(/-draft$/)) {
         resolve.push(draft);
       }
-      if (!live.workflowLocale.match(/\-draft$/)) {
+      if (!live.workflowLocale.match(/-draft$/)) {
         resolve.push(live);
       }
       // We're going in this direction for visual diff ONLY
@@ -982,7 +982,7 @@ module.exports = function(self, options) {
         return self.resolveRelationships(req, version, self.draftify(version.workflowLocale), callback);
       }, callback);
     }
-    
+
     function generateDiff(callback) {
       self.deleteExcludedProperties(live);
       self.deleteExcludedProperties(draft);
@@ -995,7 +995,7 @@ module.exports = function(self, options) {
     var slug = req.query.slug;
     var locale = req.query.locale;
     var criteria = {
-      slug: slug,
+      slug: slug
     };
     if (_.has(self.locales, locale)) {
       req.locale = locale;
@@ -1006,6 +1006,10 @@ module.exports = function(self, options) {
       return self.acceptCrossDomainSessionToken(req);
     }
     return self.apos.docs.find(req, criteria).published(null).toObject(function(err, doc) {
+      if (err) {
+        console.error(err);
+        return res.status(500).send('error');
+      }
       if (doc._url) {
         // Last step is to use the workflowLocale query parameter to
         // disambiguate cases where locales have the same URL, which
@@ -1016,9 +1020,8 @@ module.exports = function(self, options) {
       }
     });
   });
-  
-  self.route('post', 'accept-cross-domain-session-token', function(req, res) {
 
+  self.route('post', 'accept-cross-domain-session-token', function(req, res) {
 
   });
 };
