+++ lib/tasks.js
@@ -22,7 +22,7 @@ module.exports = function(self, options) {
       self.removeNumberedParkedPagesTask
     );
   };
-  
+
   self.enableResolveJoinIdsTask = function() {
     self.apos.tasks.add(self.__meta.name, 'resolve-join-ids',
       'One-time fix for join ids not pointing to the right locale',
@@ -32,7 +32,7 @@ module.exports = function(self, options) {
 
   self.addMissingLocalesTask = function(apos, argv, callback) {
     var req = self.apos.tasks.getReq();
-    
+
     return async.series([
       fixIndexes,
       noLocales,
@@ -42,7 +42,7 @@ module.exports = function(self, options) {
     ], function(err) {
       return callback(err);
     });
-    
+
     function fixIndexes(callback) {
       var old;
       return async.series([
@@ -72,7 +72,7 @@ module.exports = function(self, options) {
         return self.apos.docs.db.ensureIndex({ path: 1, workflowLocaleForPathIndex: 1 }, { unique: true, sparse: true }, callback);
       }
       function dropOldSlug(callback) {
-        var existing =_.find(old, function(index) {
+        var existing = _.find(old, function(index) {
           return index.key && index.key.slug && (!index.key.workflowLocale);
         });
         if (!existing) {
@@ -81,7 +81,7 @@ module.exports = function(self, options) {
         return self.apos.docs.db.dropIndex(existing.name, callback);
       }
       function dropOldPath(callback) {
-        var existing =_.find(old, function(index) {
+        var existing = _.find(old, function(index) {
           return index.key && index.key.path && (!index.key.workflowLocaleForPathIndex);
         });
         if (!existing) {
@@ -90,7 +90,7 @@ module.exports = function(self, options) {
         return self.apos.docs.db.dropIndex(existing.name, callback);
       }
     }
-    
+
     function noLocales(callback) {
       return self.apos.migrations.eachDoc({ workflowLocale: { $exists: 0 } }, function(doc, callback) {
         if (!self.includeType(doc.type)) {
@@ -103,7 +103,7 @@ module.exports = function(self, options) {
         return self.apos.docs.update(req, doc, { workflowMissingLocalesLive: argv.live ? true : 'liveOnly' }, callback);
       }, callback);
     }
-    
+
     function missingSomeLocales(callback) {
       return self.apos.migrations.eachDoc({ workflowLocale: self.defaultLocale }, function(doc, callback) {
         if (!self.includeType(doc.type)) {
@@ -115,11 +115,11 @@ module.exports = function(self, options) {
         });
       }, callback);
     }
-    
+
     function resolve(callback) {
-      return self.resolveDeferredRelationships(callback);      
+      return self.resolveDeferredRelationships(callback);
     }
-    
+
     function fixPermissions(callback) {
       var fields = [ 'viewUsersIds', 'viewGroupsIds',
         'editUsersIds', 'editGroupsIds',
@@ -141,15 +141,14 @@ module.exports = function(self, options) {
   };
 
   self.resolveJoinIdsTask = function(apos, argv, callback) {
-    var req = self.apos.tasks.getReq();
-    
+
     return async.series([
       update,
       resolve
     ], function(err) {
       return callback(err);
     });
-    
+
     function update(callback) {
       return self.apos.docs.db.update({
         workflowLocale: { $exists: 1 }
@@ -163,13 +162,13 @@ module.exports = function(self, options) {
       },
       callback);
     }
-    
+
     function resolve(callback) {
       return self.resolveDeferredRelationships(callback);
     }
 
   };
-  
+
   self.addLocalePrefixesTask = function(apos, argv, callback) {
     var req = self.apos.tasks.getReq();
     return self.apos.migrations.eachDoc({ slug: /^\// }, function(page, callback) {
@@ -179,7 +178,7 @@ module.exports = function(self, options) {
       return self.apos.pages.update(req, page, { permissions: false }, callback);
     }, callback);
   };
-  
+
   self.removeNumberedParkedPagesTask = function(apos, argv, callback) {
     return self.apos.docs.db.remove({
       parked: true,
