+++ package.json
@@ -5,16 +5,24 @@
   "main": "index.js",
   "dependencies": {
     "async": "^1.5.2",
+    "bluebird": "^3.5.1",
     "deep-get-set": "^1.1.0",
     "jsondiffpatch": "^0.2.5",
     "lodash": "^3.10.1"
   },
   "devDependencies": {
     "apostrophe": "^2.43.0",
-    "mocha": "^3.5.3"
+    "mocha": "^3.5.3",
+    "eslint": "^4.15.0",
+    "eslint-config-punkave": "^1.0.10",
+    "eslint-config-standard": "^11.0.0-beta.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-standard": "^3.0.1"
   },
   "scripts": {
-    "test": "mocha test/test.js"
+    "test": "mocha test/test1.js && mocha test/test2.js && eslint ."
   },
   "repository": {
     "type": "git",
