+++ public/js/force-export-widget-modal.js
@@ -5,7 +5,7 @@ apos.define('apostrophe-workflow-force-export-widget-modal', {
   extend: 'apostrophe-workflow-export-modal',
 
   source: 'force-export-widget-modal',
-  
+
   verb: 'force-export-widget',
 
   construct: function(self, options) {
@@ -14,7 +14,7 @@ apos.define('apostrophe-workflow-force-export-widget-modal', {
       return setImmediate(callback);
     };
   }
-  
+
   // The base class already sends everything in options.body so no
   // further modifications are needed on the front end so far
